+++ src/core/grid_based_algorithms/lbgpu.cpp
@@ -132,18 +132,18 @@ void lb_reinit_fluid_gpu() {
  *  See @cite dunweg07a and @cite dhumieres09a.
  */
 void lb_reinit_parameters_gpu() {
-  lbpar_gpu.time_step = (float)time_step;
+  lbpar_gpu.time_step = static_cast<float>(time_step);
   lbpar_gpu.mu = 0.0;
 
   if (lbpar_gpu.viscosity > 0.0 && lbpar_gpu.agrid > 0.0 &&
       lbpar_gpu.tau > 0.0) {
     /* Eq. (80) @cite dunweg07a. */
-    lbpar_gpu.gamma_shear = 1. - 2. / (6. * lbpar_gpu.viscosity + 1.);
+    lbpar_gpu.gamma_shear = 1.f - 2.f / (6.f * lbpar_gpu.viscosity + 1.f);
   }
 
   if (lbpar_gpu.bulk_viscosity > 0.0) {
     /* Eq. (81) @cite dunweg07a. */
-    lbpar_gpu.gamma_bulk = 1. - 2. / (9. * lbpar_gpu.bulk_viscosity + 1.);
+    lbpar_gpu.gamma_bulk = 1.f - 2.f / (9.f * lbpar_gpu.bulk_viscosity + 1.f);
   }
 
   // By default, gamma_even and gamma_odd are chosen such that the MRT becomes
@@ -173,13 +173,12 @@ void lb_reinit_parameters_gpu() {
 
 #ifdef ELECTROKINETICS
   if (ek_initialized) {
-    lbpar_gpu.dim_x = (unsigned int)round(
-        box_geo.length()[0] /
-        lbpar_gpu.agrid); // TODO code duplication with lb.c start
+    lbpar_gpu.dim_x =
+        static_cast<unsigned int>(round(box_geo.length()[0] / lbpar_gpu.agrid)); // TODO code duplication with lb.c start
     lbpar_gpu.dim_y =
-        (unsigned int)round(box_geo.length()[1] / lbpar_gpu.agrid);
+        static_cast<unsigned int>(round(box_geo.length()[1] / lbpar_gpu.agrid));
     lbpar_gpu.dim_z =
-        (unsigned int)round(box_geo.length()[2] / lbpar_gpu.agrid);
+        static_cast<unsigned int>(round(box_geo.length()[2] / lbpar_gpu.agrid));
 
     unsigned int tmp[3];
 
@@ -192,7 +191,8 @@ void lb_reinit_parameters_gpu() {
 
     for (dir = 0; dir < 3; dir++) {
       /* check if box_l is compatible with lattice spacing */
-      if (fabs(box_geo.length()[dir] - tmp[dir] * lbpar_gpu.agrid) > 1.0e-3) {
+      if (fabs(box_geo.length()[dir] -
+               static_cast<float>(tmp[dir]) * lbpar_gpu.agrid) > 1.0e-3) {
         runtimeErrorMsg() << "Lattice spacing lbpar_gpu.agrid= "
                           << lbpar_gpu.agrid << " is incompatible with box_l["
                           << dir << "]=" << box_geo.length()[dir];
@@ -201,7 +201,7 @@ void lb_reinit_parameters_gpu() {
 
     lbpar_gpu.number_of_nodes =
         lbpar_gpu.dim_x * lbpar_gpu.dim_y * lbpar_gpu.dim_z;
-    lbpar_gpu.tau = (float)time_step; // TODO code duplication with lb.c end
+    lbpar_gpu.tau = static_cast<float>(time_step); // TODO code duplication with lb.c end
   }
 #endif
 
@@ -254,7 +254,8 @@ void lb_set_agrid_gpu(double agrid) {
   /* sanity checks */
   for (int dir = 0; dir < 3; dir++) {
     /* check if box_l is compatible with lattice spacing */
-    if (fabs(box_geo.length()[dir] - tmp[dir] * agrid) > ROUND_ERROR_PREC) {
+    if (fabs(box_geo.length()[dir] - static_cast<float>(tmp[dir]) * agrid) >
+        ROUND_ERROR_PREC) {
       runtimeErrorMsg() << "Lattice spacing p_agrid= " << agrid
                         << " is incompatible with box_l[" << dir
                         << "]=" << box_geo.length()[dir]
