+++ src/core/statistics.cpp
@@ -338,7 +338,7 @@ void calc_rdf(PartCfg &partCfg, int const *p1_types, int n_p1,
     auto const r_out = r_in + bin_width;
     auto const bin_volume = (4.0 / 3.0) * Utils::pi() *
                             ((r_out * r_out * r_out) - (r_in * r_in * r_in));
-    rdf[i] *= volume / (bin_volume * cnt);
+    rdf[i] *= volume / (bin_volume * static_cast<double>(cnt));
   }
 }
 
@@ -410,7 +410,7 @@ void calc_rdf_av(PartCfg &partCfg, int const *p1_types, int n_p1,
       auto const r_out = r_in + bin_width;
       auto const bin_volume = (4.0 / 3.0) * Utils::pi() *
                               ((r_out * r_out * r_out) - (r_in * r_in * r_in));
-      rdf[i] += rdf_tmp[i] * volume / (bin_volume * cnt);
+      rdf[i] += rdf_tmp[i] * volume / (bin_volume * static_cast<double>(cnt));
     }
 
     cnt_conf++;
@@ -524,8 +524,9 @@ void obsstat_realloc_and_clear(Observable_stat *stat, int n_pre, int n_bonded,
 
   // Number of doubles to store pressure in
   const int total =
-      c_size * (n_pre + bonded_ia_params.size() + n_non_bonded + n_coulomb +
-                n_dipolar + n_vs + Observable_stat::n_external_field);
+      c_size *
+      (n_pre + static_cast<int>(bonded_ia_params.size()) + n_non_bonded +
+       n_coulomb + n_dipolar + n_vs + Observable_stat::n_external_field);
 
   // Allocate mem for the double list
   stat->data.resize(total);
